requirements.txt
fastapi
uvicorn
pillow
transformers
torch
diffusers
